Create README.md
# README
## README file for Final Exam Christine Ottinger

### For this project I decided to submit both a Jupyter Notebook and my zipped .tar.gz file just incase I did not do my .tar.gz correctly.

**To open my .tar.gz file please run the following code in terminal:**
  * conda install py37
  * cd 'Final Exam Ottinger'.ipynb
  * jupyter notebook


### *If this does NOT work.. Please refer to the Jupyter Notebook and run all*


For my project I decided to look into what should be recommended for a **'WOOD 2 DRAWER CABINET WHITE FINISH'.** 
My output is as followed with both the product and score: 

  * WHITE METAL LANTERN   12.682215743440233
  * GLASS STAR FROSTED T-LIGHT HOLDER   12.682215743440233
  * VINTAGE BILLBOARD LOVE/HATE MUG   12.448979591836737
  * SAVE THE PLANET MUG   10.255102040816327
  * RETRO COFFEE MUGS ASSORTED   10.255102040816327
  * WOOD S/3 CABINET ANT WHITE FINISH   9.959183673469385
  * WOODEN FRAME ANTIQUE WHITE    9.959183673469385
  * WOODEN PICTURE FRAME WHITE FINISH   9.959183673469385
  * KNITTED UNION FLAG HOT WATER BOTTLE   9.6734693877551
  * VINTAGE BILLBOARD DRINK ME MUG   8.68804664723032


### Kibana
I did have some struggles with Kibana, and with the time constraint I did not have the time to fully explore all the options of visualizations: 
I did want to add a visual to show I did attempt this aspect of the assignment. I decided to look into the countries described in this dataset 
and see if there were any neighboring countries included in the dataset. You can see that there are some inputs from France, Norway and the Netherlands.

Here is the visual output if needed: 
![GitHub Logo](/home/vagrant/work/Final Exam/Capture.png)
Format: ![Alt Text](url)